feat: Add scan grocery page and camera component
