+++ src/app/scanGrocery/page.jsx
@@ -1,14 +1,10 @@
-"use client";
-
-import { Input } from "@/components/ui/input";
-import AddItems from "./form";
-
 export default function ScanGrocery() {
   return (
     <main>
       {/* <Input placeholder="Хүнсний нэр" /> */}
       <h1>Scan Grocery</h1>
-      <AddItems />
+
+      {/* <AddItems /> */}
 
       {/* <CameraComponent /> */}
     </main>
